requirements.txt
Django==1.4.5
django-decouple==1.1
dj-database-url==0.2.1
unipath==1.0
xlwt==0.7.5
Babel==0.9.6
django-tinymce==1.5.1
django-sekizai==0.7
South==0.7.6
django-menus==1.1.1
django-mptt==0.5.2
django-cms==2.3.3
django-appmedia==1.0.1
MySQL-python==1.2.3
PIL==1.1.7
beautifulsoup==3.2.1
Fabric==1.6.1
sphinx
django-reversion
cuisine==0.6.4
model-mommy==1.0
tablib==0.9.11
Pygments==1.6
sqlparse==0.1.8
django-aggregate-if==0.3.1
